+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Add internal command `update-formula-tests` to automatically update formula test expectations
 * Add program change log in server mode to track program updates
 * Use v2 API endpoints to download OEIS data and report stats
+* Set user agent in web requests
 
 ## v25.12.1